Disable task retries for the moment.
Problem that caused the exception needs to be inspected to determine whether
the task can be retried or not. For example, problem that is caused by an
invalid control files must not be retried. On the other hand, network problem
that prevents the builder to download required packages can be retried because
it has nothing to do with package error.